handle build when requirements.txt file is missing

+++ ironbank/build_ironbank_docker.sh
@@ -125,7 +125,11 @@ function upload_image_to_artifacts {
   SOURCE_PATH="ironbank/$IMAGE_NAME"
   cp -r $SOURCE_PATH $TARGET_PATH
   cp $CURRENT_DIR/docker/$IMAGE_NAME/requirements.txt $TARGET_PATH
-  rm $SOURCE_PATH/docker_packages_metadata.txt
+  if [[ -f $SOURCE_PATH/docker_packages_metadata.txt ]] && [[ -s $SOURCE_PATH/docker_packages_metadata.txt ]]; then
+    rm $SOURCE_PATH/docker_packages_metadata.txt
+    echo "The file docker_packages_metadata.txt has been deleted."
+  fi
+
 }
 
 # $1: docker image dir (~/../docker/$IMAGE_NAME)
